Revert "test(hipfile): Add tests that modify data on the GPU"
