+++ projects/hipfile/test/CMakeLists.txt
@@ -36,11 +36,7 @@ if(AIS_BUILD_NVIDIA_DETAIL)
     list(APPEND UNIT_TEST_SOURCE_FILES nvidia_detail/cufile-api-compat.cpp)
     list(APPEND TEST_SYSINCLS ${HIPFILE_NVIDIA_SOURCE_PATH})
 else()
-    list(APPEND SYSTEM_TEST_SOURCE_FILES
-        system/amd/io.cpp
-        system/amd/io-data-modification.cpp
-        system/amd/io-data-modification-kernel.hip
-    )
+    list(APPEND SYSTEM_TEST_SOURCE_FILES system/amd/io.cpp)
     list(APPEND SHARED_SOURCE_FILES common/ais-capability.cpp)
     list(APPEND TEST_SYSINCLS ${HIPFILE_AMD_SOURCE_PATH})
 endif()
