+++ projects/hipfile/test/system/amd/io.cpp
@@ -9,7 +9,7 @@
 #include "hipfile-warnings.h"
 #include "hipfile.h"
 
-#include "io-test.h"
+#include "ais-capability.h"
 #include "test-common.h"
 #include "test-options.h"
 
@@ -29,7 +29,46 @@ extern SystemTestOptions test_env;
 
 using namespace hipFile;
 
+namespace {
+
+// Gate fastpath-only tests on AIS capability.
+void
+enforceFastpathGate(hipFileHandle_t handle, void *device_buffer)
+{
+    hipFile::test::AisCapability ais_capability{test_env.allow_skip_fastpath};
+
+    const auto decision = ais_capability.populate(handle, device_buffer);
+
+    if (decision == hipFile::test::AisCapability::GateDecision::Run) {
+        return;
+    }
+
+    if (decision == hipFile::test::AisCapability::GateDecision::Skip) {
+        // Keep this marker synchronized with test/CMakeLists.txt SKIP_REGULAR_EXPRESSION.
+        GTEST_SKIP() << "fastpath not available in this environment\n" << ais_capability.report();
+    }
+
+    FAIL() << "Fastpath Validation Failed!\n" << ais_capability.report() << "\n" << ais_capability.skipHint();
+}
+
+}
+
 HIPFILE_WARN_NO_GLOBAL_CTOR_OFF
+
+enum class IoTestBackend {
+    Fastpath,
+    Fallback,
+};
+
+struct IoTestParam {
+    IoTestBackend backend;
+    std::string   name;
+};
+
+HIPFILE_WARN_NO_EXIT_DTOR_OFF
+static std::array<IoTestParam, 2> io_test_params{
+    {{IoTestBackend::Fastpath, "Fastpath"}, {IoTestBackend::Fallback, "Fallback"}}};
+HIPFILE_WARN_NO_EXIT_DTOR_ON
 
 struct HipFileIo : public testing::TestWithParam<IoTestParam> {
